ExploadWar fails when CWD is not project dir. Run from IDE fail if not configured to run plugin tests in plugin root directory.
java.lang.Exception: Failed to initialize exploded war
	at org.jvnet.hudson.test.WarExploder.getExplodedDir(WarExploder.java:49)
	at org.jvnet.hudson.test.JenkinsRule.createWebServer(JenkinsRule.java:582)
	at org.jvnet.hudson.test.JenkinsRule.newHudson(JenkinsRule.java:538)
	at org.jvnet.hudson.test.JenkinsRule.before(JenkinsRule.java:332)
	at org.jvnet.hudson.test.JenkinsRule$2.evaluate(JenkinsRule.java:482)
	at org.junit.rules.ExternalResource$1.evaluate(ExternalResource.java:48)
	at org.junit.rules.RunRules.evaluate(RunRules.java:20)
	at org.junit.runners.ParentRunner.runLeaf(ParentRunner.java:271)
	at org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:70)
	at org.junit.runners.BlockJUnit4ClassRunner.runChild(BlockJUnit4ClassRunner.java:50)
	at org.junit.runners.ParentRunner$3.run(ParentRunner.java:238)
	at org.junit.runners.ParentRunner$1.schedule(ParentRunner.java:63)
	at org.junit.runners.ParentRunner.runChildren(ParentRunner.java:236)
	at org.junit.runners.ParentRunner.access$000(ParentRunner.java:53)
	at org.junit.runners.ParentRunner$2.evaluate(ParentRunner.java:229)
	at org.junit.runners.ParentRunner.run(ParentRunner.java:309)
	at org.junit.runner.JUnitCore.run(JUnitCore.java:160)
	at com.intellij.junit4.JUnit4IdeaTestRunner.startRunnerWithArgs(JUnit4IdeaTestRunner.java:74)
	at com.intellij.rt.execution.junit.JUnitStarter.prepareStreamsAndStart(JUnitStarter.java:211)
	at com.intellij.rt.execution.junit.JUnitStarter.main(JUnitStarter.java:67)
	at sun.reflect.NativeMethodAccessorImpl.invoke0(Native Method)
	at sun.reflect.NativeMethodAccessorImpl.invoke(NativeMethodAccessorImpl.java:57)
	at com.intellij.rt.execution.application.AppMain.main(AppMain.java:134)
Caused by: java.io.FileNotFoundException: /Users/integer/Documents/idea/jenkins-ws-git-plugin/./target/jenkins-for-test.exploding (No such file or directory)
	at java.io.FileOutputStream.open(Native Method)
	at java.io.FileOutputStream.<init>(FileOutputStream.java:212)
	at java.io.FileOutputStream.<init>(FileOutputStream.java:104)
	at org.jvnet.hudson.test.WarExploder.explode(WarExploder.java:99)
	at org.jvnet.hudson.test.WarExploder.<clinit>(WarExploder.java:58)
	... 24 more

+++ test/src/main/java/org/jvnet/hudson/test/WarExploder.java
@@ -87,7 +87,6 @@ final class WarExploder {
             throw new AssertionError("jenkins.war is not in the classpath. If you are running this from the core workspace, run 'mvn install' to create the war image in war/target/jenkins");
         File war = Which.jarFile(Class.forName("executable.Executable"));
 
-        // TODO this assumes that the CWD of the Maven process is the plugin ${basedir}, which may not be the case
         File explodeDir = new File("./target/jenkins-for-test").getAbsoluteFile();
         while (new File(explodeDir + ".exploding").isFile()) {
             explodeDir = new File(explodeDir + "x");
@@ -96,8 +95,12 @@ final class WarExploder {
 
         if(!timestamp.exists() || (timestamp.lastModified()!=war.lastModified())) {
             System.out.println("Exploding " + war + " into " + explodeDir);
-            new FileOutputStream(explodeDir + ".exploding").close();
-            new FilePath(explodeDir).deleteRecursive();
+			if (new File(explodeDir + ".exploding").exists()){
+				new FilePath(explodeDir).deleteRecursive();
+			}
+			if (!explodeDir.exists()) {
+				explodeDir.mkdirs();
+			}
             new FilePath(war).unzip(new FilePath(explodeDir));
             if(!explodeDir.exists())    // this is supposed to be impossible, but I'm investigating HUDSON-2605
                 throw new IOException("Failed to explode "+war);
